Moved static.c to base.c
